tests work correctly
