+++ docker-compose.yml
@@ -14,9 +14,8 @@ services:
       context: .
       dockerfile: Dockerfile
     command: >
-      sh -c "/wait-for-postgres.sh db &&
-             python manage.py migrate &&
-             gunicorn config.wsgi:application --bind 0.0.0.0:8000"
+      /wait-for-postgres.sh db python manage.py migrate &&
+      gunicorn config.wsgi:application --bind 0.0.0.0:8000
     ports:
       - "8000:8000"
     volumes:
@@ -25,6 +24,8 @@ services:
       - db
     env_file:
       - .env
+    environment:
+      - POSTGRES_HOST=db
     networks:
       - shopnet
